+++ scilab/modules/io/sci_gateway/cpp/sci_lib.cpp
@@ -28,7 +28,7 @@ extern "C"
 }
 
 /*--------------------------------------------------------------------------*/
-types::Function::ReturnValue sci_lib(types::typed_list &in, int /*_iRetCount*/, types::typed_list &out)
+types::Function::ReturnValue sci_gwlib(types::typed_list &in, int /*_iRetCount*/, types::typed_list &out)
 {
     if (in.size() != 1)
     {